+++ angular/src/directives/navigation/stack-controller.ts
@@ -273,7 +273,6 @@ export class StackController {
 
       if ((containerEl as any).commit) {
         return containerEl.commit(enteringEl, leavingEl, {
-          deepWait: true,
           duration: direction === undefined ? 0 : undefined,
           direction,
           showGoBack,