+++ core/src/components/datetime/test/minmax/datetime.e2e.ts
@@ -257,9 +257,7 @@ test.describe('datetime: minmax', () => {
 
       await ionChange.next();
 
-      const value = await datetime.evaluate((el: HTMLIonDatetimeElement) => el.value);
-      await expect(typeof value).toBe('string');
-      await expect(value!.includes('2022-10-10T08:00')).toBe(true);
+      await expect(datetime).toHaveJSProperty('value', '2022-10-10T08:00:00');
     });
     test('should reset to max time if out of bounds', async ({ page }) => {
       await page.setContent(`
@@ -278,9 +276,7 @@ test.describe('datetime: minmax', () => {
 
       await ionChange.next();
 
-      const value = await datetime.evaluate((el: HTMLIonDatetimeElement) => el.value);
-      await expect(typeof value).toBe('string');
-      await expect(value!.includes('2022-10-10T08:00')).toBe(true);
+      await expect(datetime).toHaveJSProperty('value', '2022-10-10T08:00:00');
     });
   });