+++ core/src/components/tab-bar/tab-bar.tsx
@@ -2,10 +2,12 @@ import type { ComponentInterface, EventEmitter } from '@stencil/core';
 import { Component, Element, Event, Host, Prop, State, Watch, h } from '@stencil/core';
 
 import { getIonMode } from '../../global/ionic-global';
-import type { Color, TabBarChangedEventDetail } from '../../interface';
+import type { Color } from '../../interface';
 import type { KeyboardController } from '../../utils/keyboard/keyboard-controller';
 import { createKeyboardController } from '../../utils/keyboard/keyboard-controller';
 import { createColorClasses } from '../../utils/theme';
+
+import type { TabBarChangedEventDetail } from './tab-bar-interface';
 
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines which platform styles to use.