+++ docs/CHANGELOG.md
@@ -3,7 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.5...v7.0.0) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-rc.5](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.4...v7.0.0-rc.5) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.7.1](https://github.com/ionic-team/ionic-docs/compare/v6.7.0...v6.7.1) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-rc.4](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.3...v7.0.0-rc.4) (2023-03-27)
 
 **Note:** Version bump only for package @ionic/docs
 
@@ -19,7 +43,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-rc.3](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.2...v7.0.0-rc.3) (2023-03-22)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.6.3](https://github.com/ionic-team/ionic-docs/compare/v6.6.2...v6.6.3) (2023-03-22)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-rc.2](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.1...v7.0.0-rc.2) (2023-03-15)
 
 **Note:** Version bump only for package @ionic/docs
 
@@ -35,7 +75,32 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-rc.1](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-rc.0...v7.0.0-rc.1) (2023-03-08)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.6.1](https://github.com/ionic-team/ionic-docs/compare/v6.6.0...v6.6.1) (2023-03-08)
+
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-rc.0](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.6...v7.0.0-rc.0) (2023-03-01)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-beta.6](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.5...v7.0.0-beta.6) (2023-03-01)
 
 **Note:** Version bump only for package @ionic/docs
 
@@ -59,6 +124,22 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-beta.5](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.4...v7.0.0-beta.5) (2023-02-27)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-beta.4](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.3...v7.0.0-beta.4) (2023-02-22)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.5.6](https://github.com/ionic-team/ionic-docs/compare/v6.5.5...v6.5.6) (2023-02-22)
 
 **Note:** Version bump only for package @ionic/docs
@@ -75,7 +156,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-beta.3](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.2...v7.0.0-beta.3) (2023-02-15)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.5.4](https://github.com/ionic-team/ionic-docs/compare/v6.5.3...v6.5.4) (2023-02-15)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-beta.2](https://github.com/ionic-team/ionic-docs/compare/v7.0.0-beta.1...v7.0.0-beta.2) (2023-02-08)
 
 **Note:** Version bump only for package @ionic/docs
 
@@ -91,7 +188,24 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-beta.1](https://github.com/ionic-team/ionic-docs/compare/v6.5.2...v7.0.0-beta.1) (2023-02-01)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
 ## [6.5.2](https://github.com/ionic-team/ionic-docs/compare/v6.5.1...v6.5.2) (2023-02-01)
+
+**Note:** Version bump only for package @ionic/docs
+
+
+
+
+
+# [7.0.0-beta.0](https://github.com/ionic-team/ionic-docs/compare/v6.5.1...v7.0.0-beta.0) (2023-01-25)
+
 
 **Note:** Version bump only for package @ionic/docs